+++ .gitignore
@@ -1,5 +1,5 @@
 # Project files to exclude.
-
+config.json
 
 # WIP
 sysdvr/
