+++ config.json
@@ -1,17 +1,5 @@
 {
-    "DiscordBotPrefix": "$",
-    "DiscordToken": "",
-    "EnableDebugLogging": "False",
-    "EventNotificationChannelIdForShiny": "",
-    "EventNotificationChannelIdForBrilliant": "",
-    "EventNotificationChannelIdForMark": "",
-    "EventNotificationChannelIdForRareMark": "",
-    "IP": "192.168.1.63",
-	"USB": false,
-	"SeedCheckOnly": false,
-    "NotificationChannelForInfo": "",
-    "TextNotificationChannelIdForShiny": "",
-    "TextNotificationChannelIdForBrilliant": "",
-    "TextNotificationChannelIdForMark": "",
-    "TextNotificationChannelIdForRareMark": ""
+  "IP": "192.168.178.37",
+  "USB": false,
+  "SeedCheckOnly": false
 }
