--- templates/base.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="color-scheme" content="dark light">
-    <title>PLA {% if title %}{{ title }}{% else %}Multi Checker{% endif %}</title>
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Lato&display=swap" rel="stylesheet">
-    <link href="{{ url_for('static', filename='css/style.css') }}" rel="stylesheet">
-    {% block head %}
-    {% endblock %}
-</head>
-
-<body>
-    <header class="pla-header space-between">
-        <h1 class="pla-heading"><span class="pla-heading-highlight">PLA</span> {% if title %}{{ title }}{% else %}Multi Checker{% endif %}</h1>
-
-        <button class="theme-toggle" id="theme-toggle" title="Toggles light and dark mode" aria-live="light">
-            <svg aria-hidden="true" width="24" height="24" viewBox="0 0 24 24">
-                <g class="theme-toggle-light">
-                    <path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z"></path>
-                </g>
-                <g class="theme-toggle-dark">
-                    <circle cx="12" cy="12" r="5"></circle>
-                    <line x1="12" y1="1" x2="12" y2="3"></line>
-                    <line x1="12" y1="21" x2="12" y2="23"></line>
-                    <line x1="4.22" y1="4.22" x2="5.64" y2="5.64"></line>
-                    <line x1="18.36" y1="18.36" x2="19.78" y2="19.78"></line>
-                    <line x1="1" y1="12" x2="3" y2="12"></line>
-                    <line x1="21" y1="12" x2="23" y2="12"></line>
-                    <line x1="4.22" y1="19.78" x2="5.64" y2="18.36"></line>
-                    <line x1="18.36" y1="5.64" x2="19.78" y2="4.22"></line>
-                </g>
-            </svg>
-        </button>
-    </header>
-
-    {% if web %}
-    {% include 'nav/nav-web.html' %}
-    {% else %}
-    {% include 'nav/nav-python.html' %}
-    {% endif %}
-    
-    <aside class="pla-sidebar">
-        {% block sidebar %}
-        {% endblock %}
-    </aside>
-
-    <main class="pla-main">
-        {% block main %}
-        {% endblock %}
-    </main>
-
-    <footer class="pla-footer">
-        <a href="https://discord.gg/2JSZnm8GHf" target="_blank">Discord</a>
-        <a href="https://www.youtube.com/channel/UCrItwn5zVd2jinwkznYn-dQ" target="_blank">YouTube Guides</a>
-        <a href="https://github.com/capitalism-sudo/pla-multi-checker/wiki" target="_blank">Wiki</a>
-    </footer>
-
-    <div class="pla-background-1"></div>
-    <div class="pla-background-2"></div>
-
-    {% block templates %}
-    {% endblock %}
-
-    <script src="{{ url_for('static', filename='js/pla-common.js') }}"></script>
-    {% block script %}
-    {% endblock %}
-</body>
-
-</html>
