+++ templates/index.html
@@ -1,23 +1,7 @@
-{% extends "base.html" %}
+{% extends "base/base.html" %}
 
 {% block main %}
     <section class="pla-section-results">
         Welcome! Special Thanks to Lincoln, Zyro, Santacrab420, Kurt and Anubis for code, insights and help!
-    </section>
-{% endblock %}
-
-{% block sidebar %}
-    <section class="pla-section-options">
-        <div class="pla-control-actions">
-            <a class="pla-button pla-button-action" href="/mmos">PLA MMO Checker (CFW NEEDED)</a>
-        </div>
-
-        <div class="pla-control-actions">
-            <a class="pla-button pla-button-action" href="/seed">PLA Seed Checker (No CFW Needed)</a>
-        </div>
-        
-        <div class="pla-control-actions">
-            <a class="pla-button pla-button-action" href="/distortion">PLA Distortion Checker (CFW NEEDED)</a>
-        </div>
     </section>
 {% endblock %}
