+++ templates/pages/distortions.html
@@ -1,9 +1,13 @@
-{% extends "base.html" %}
+{% extends "base/base-sidebar.html" %}
+
+{% block title %}Distortions Checker{% endblock %}
 
 {% block main %}
     <section class="pla-section-results">
+        <div data-pla-messages></div>
         <div class="pla-results" data-pla-results>
-            Search to begin
+            <p class="pla-results-message">Welcome to the distortion checker. Choose a map and begin hunting distortion
+                Pokémon!</p>
         </div>
     </section>
 {% endblock %}
@@ -48,8 +52,10 @@
         </div>
 
         <div class="pla-control-actions">
-            <button type="button" class="pla-button pla-button-action" onclick="checkDistortions()">Check Distortions</button>
-            <button type="button" class="pla-button pla-button-action" onclick="createDistortion()">Create Distortion</button>
+            <button type="button" class="pla-button pla-button-action" id="pla-button-check-distortions">Check
+                Distortions</button>
+            <button type="button" class="pla-button pla-button-action" id="pla-button-create-distortion">Create
+                Distortion</button>
         </div>
     </section>
 
@@ -81,6 +87,7 @@
             <ul class="pla-results-ul">
                 <li><span class="pla-results-label">Nature:</span> <span data-pla-results-nature></span></li>
                 <li><span class="pla-results-label">Gender:</span> <span data-pla-results-gender></span></li>
+                <li><span class="pla-results-label">Rolls:</span> <span data-pla-results-rolls></span></li>
             </ul>
 
             <ul class="pla-results-ul">
@@ -102,21 +109,8 @@
             </div>
         </div>
     </template>
-
-    <template data-pla-spinner>
-        <div class="lds-roller">
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-        </div>
-    </template>
 {% endblock %}
 
 {% block script %}
-    <script src="{{ url_for('static', filename='js/pla-distortion-checker.js') }}"></script>
+    <script type="module" src="{{ url_for('static', filename='js/distortions.js') }}"></script>
 {% endblock %}
