+++ templates/pages/spawns.html
@@ -1,4 +1,6 @@
-{% extends "base.html" %}
+{% extends "base/base-sidebar.html" %}
+
+{% block title %}Spawn Checker{% endblock %}
 
 {% block head %}
 	<script type="text/javascript" src="{{ url_for('static', filename='js/jquery.min.js') }}"></script>
@@ -63,9 +65,10 @@
 {% endblock %}
 
 {% block main %}
-	<section class="pla-section-results flow">
+	<section class="pla-section-results">
+		<div data-pla-messages></div>
 		<div class="pla-results" data-pla-results>
-			This tool was created by Cappy, with help from Anubis, Kurt, Lincoln, Zyro and Santacrab420 for code and other assistance.
+			<p class="pla-results-message">This tool was created by Cappy, with help from Anubis, Kurt, Lincoln, Zyro and Santacrab420 for code and other assistance.</p>
 		</div>
 	</section>
 {% endblock %}
@@ -74,7 +77,7 @@
 	<section class="pla-section-options">
 		<div class="pla-control">
 			<label for="inputseed">Seed</label>
-			<input type="text" id="inputseed" placeholder="Input Seed">
+			<input type="number" id="inputseed" placeholder="Input Seed">
 		</div>
 
 		<div class="pla-control">
@@ -112,7 +115,7 @@
 		</div>
 
 		<div class="pla-control-actions">
-			<button class="pla-button pla-button-action" onclick="checkalphaadv()">Find Alpha Shiny Advances</button>
+			<button class="pla-button pla-button-action" id="pla-button-checkalphaadv">Find Alpha Shiny Advances</button>
 		</div>
 	</section>
 
@@ -176,6 +179,7 @@
             <ul class="pla-results-ul">
                 <li><span class="pla-results-label">Nature:</span> <span data-pla-results-nature></span></li>
                 <li><span class="pla-results-label">Gender:</span> <span data-pla-results-gender></span></li>
+				<li><span class="pla-results-label">Rolls:</span> <span data-pla-results-rolls></span></li>
             </ul>
 
             <div class="pla-results-ivs-container">
@@ -191,21 +195,8 @@
             </div>
         </div>
     </template>
-
-    <template data-pla-spinner>
-        <div class="lds-roller">
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-            <div></div>
-        </div>
-    </template>
 {% endblock %}
 
 {% block script %}
-<script src="{{ url_for('static', filename='js/pla-alpha-seed-reader.js') }}"></script>
+	<script type="module" src="{{ url_for('static', filename='js/spawns.js') }}"></script>
 {% endblock %}
